requirements.txt
bcrypt
pandas
numpy
beautifulsoup4
pymysql
matplotlib
requests
streamlit_lottie
st_pages
streamlit.extras
python-dotenv
fpdf
mysql-connector-python
streamlit==1.38.0
streamlit_extras
streamlit-cookies-manager
altair

#chatbot's dependencies
tqdm
google-generativeai
streamlit
transformers
sentence-transformers
langchain
langchain-community
langsmith
tf-keras
lark-parser
qdrant_client
lark
langchain_huggingface